added documentation of central functions added new function for event handling compile fixes
git-svn-id: file:///home/scotch/motortreiber/svn/trunk/src/STM32@6 91c3317b-883a-473d-b98d-25e381635080

+++ i2c.h
@@ -22,7 +22,7 @@ struct I2C_Data {
   u8 curI2CAddr;
   u8 I2CMode;
   u8 I2CError;
-  u16 I2CErrorReason;
+  u32 I2CErrorReason;
 };
 
 extern volatile struct I2C_Data I2C1_Data;
